Move utility functions from DarpaLinks to CorefUtils, remove extraneous
